brg-6 Make a  better project structure
