--- src/App.js
@@ -1,94 +1,0 @@
-import React, { Component } from 'react';
-//import Radium, { StyleRoot } from 'radium';
-import classes from './App.module.css';
-import Person from './Person/Person';
-import ErrorBoundary from './ErrorBoundary/ErrorBoundary';
-
-class App extends Component {
-  state = {
-    persons: [
-      { id: "de3f", name: "Max", age: 28 },
-      { id: "fed4", name: "Manu", age: 29 },
-      { id: "fd4f", name: "Stephanie", age: 26 },
-    ],
-    showPersons: false
-  }
-
-  deletePersonHandler = (personIndex) => {
-    const persons = [...this.state.persons];
-    persons.splice(personIndex, 1);
-    this.setState({
-      persons: persons
-    })
-  }
-
-  nameChangedHandler = (event, id) => {
-    const personIndex = this.state.persons.findIndex(p => {
-      return p.id === id;
-    });
-
-    const person = {
-      ...this.state.persons[personIndex]
-    };
-
-    person.name = event.target.value;
-
-    const persons = [...this.state.persons];
-    persons[personIndex] = person;
-
-    this.setState({ persons: persons })
-  }
-
-  togglePersonsHandler = () => {
-    const doesShow = this.state.showPersons
-    this.setState({
-      showPersons: !doesShow
-    })
-  }
-
-  render() {
-
-    let persons = null;
-    let btnClass = '';
-
-    if (this.state.showPersons) {
-      persons = (
-        <div >
-          {this.state.persons.map((person, index) => {
-            return <ErrorBoundary key={person.id}>
-              <Person
-                name={person.name}
-                age={person.age}
-                click={this.deletePersonHandler.bind(this, index)}
-                changed={(event) => this.nameChangedHandler(event, person.id)}
-              />
-            </ErrorBoundary>
-          })}
-        </div>
-      );
-
-      btnClass = classes.Red;
-    }
-
-    const assignedClasses = [];
-    if (this.state.persons.length <= 2) {
-      assignedClasses.push(classes.red);
-    };
-    if (this.state.persons.length <= 1) {
-      assignedClasses.push(classes.bold);
-    };
-
-    return (
-      <div className={classes.App}>
-        <h1>Hi, I'm React App!</h1>
-        <p className={assignedClasses.join(' ')}>This is really working</p>
-        <button
-          className={btnClass}
-          onClick={this.togglePersonsHandler}>Toggle Persons</button>
-        {persons}
-      </div>
-    );
-  }
-}
-
-export default App;
