--- src/App.module.css
@@ -1,34 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.red {
-  color: red;
-}
-
-.bold {
-  font-weight: bold;
-}
-
-.App button {
-  border: 1px solid blue;
-  padding: 16px;
-  background-color: green;
-  font: inherit;
-  color: white;
-  cursor: pointer;
-}
-
-.App button:hover {
-  background-color: lightgreen;
-  color: black;
-}
-
-.App button.Red {
-  background-color: red;
-}
-
-.App button.Red:hover {
-  background-color: salmon;
-  color: black;
-}
