+++ src/components/Persons/Persons.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import Person from './Person/Person';
+
+const persons = (props) => props.persons.map((person, index) => {
+   return <Person
+      key={person.id}
+      name={person.name}
+      age={person.age}
+      click={props.clicked.bind(this, index)}
+      changed={(event) => props.changed(event, person.id)} />
+});
+
+export default persons;
