+++ src/containers/App.js
@@ -1,0 +1,74 @@
+import React, { Component } from 'react';
+//import Radium, { StyleRoot } from 'radium';
+import classes from './App.module.css';
+import Persons from '../components/Persons/Persons';
+import Cockpit from '../components/Cockpit/Cockpit';
+
+class App extends Component {
+   state = {
+      persons: [
+         { id: "de3f", name: "Max", age: 28 },
+         { id: "fed4", name: "Manu", age: 29 },
+         { id: "fd4f", name: "Stephanie", age: 26 },
+      ],
+      showPersons: false
+   }
+
+   deletePersonHandler = (personIndex) => {
+      const persons = [...this.state.persons];
+      persons.splice(personIndex, 1);
+      this.setState({
+         persons: persons
+      })
+   }
+
+   nameChangedHandler = (event, id) => {
+      const personIndex = this.state.persons.findIndex(p => {
+         return p.id === id;
+      });
+
+      const person = {
+         ...this.state.persons[personIndex]
+      };
+
+      person.name = event.target.value;
+
+      const persons = [...this.state.persons];
+      persons[personIndex] = person;
+
+      this.setState({ persons: persons })
+   }
+
+   togglePersonsHandler = () => {
+      const doesShow = this.state.showPersons
+      this.setState({
+         showPersons: !doesShow
+      })
+   }
+
+   render() {
+
+      let persons = null;
+
+      if (this.state.showPersons) {
+         persons = <Persons
+            persons={this.state.persons}
+            clicked={this.deletePersonHandler}
+            changed={this.nameChangedHandler} />
+      }
+
+      return (
+         <div className={classes.App}>
+            <Cockpit
+               appTitle={this.props.title}
+               persons={this.state.persons}
+               showPersons={this.state.showPersons}
+               clicked={this.togglePersonsHandler}
+            />
+            {persons}
+         </div>
+      );
+   }
+}
+
+export default App;
